requirements.txt
pywhatkit==5.4
pyinstaller==5.1
neutron-web==0.1.6
lxml==4.9.1
